fix: font_family
Comic Sans MS is not suitable for SVG files on the web. Hence using cursive instead is better for the web.

Another option is to supply multiple font families as suggested here https://stackoverflow.com/a/21335048
This would require allowing the text function to take font_family as a string or a list based on context.