Rewrite all_shapes
As reported in #149, `get_parent` was not working properly.  The
problem appears to be that `all_shapes` was not returning a complete
set of document shapes.  This commit modifies `all_shapes` to return
the union of all shapes created by Simple Inkscape Scripting and
pre-existing document shapes, which are subsequently converted to
Simple Inkscape Scripting objects.