Allow the user to ignore package-import errors
Based on Simple Inkscape Scripting's GitHub issues page, a number of
Windows and macOS users don't have NumPy or Pillow installed and have
trouble installing them in a location where Inkscape can find them.
As a workaround, the user can check a box in the dialog box to allow
those imports to fail and simply throw an exception if the user's
script accesses an unavailable feature.